images/requirements.txt
google-genai>=0.1.0
python-dotenv>=1.0.0
Pillow>=10.0.0

